[Refactor] add useReportHandler #775
