+++ components/modal/menu/ReportModal.tsx
@@ -1,9 +1,4 @@
 import { useState } from 'react';
-import { useSetRecoilState, useResetRecoilState } from 'recoil';
-import { errorState } from 'utils/recoil/error';
-import { openMenuBarState } from 'utils/recoil/layout';
-import { modalState } from 'utils/recoil/modal';
-import instance from 'utils/axios';
 import styles from 'styles/modal/menu/ReportModal.module.scss';
 
 interface Report {
@@ -11,14 +6,13 @@ interface Report {
   content: string;
 }
 
+import useReportHandler from 'hooks/modal/useReportHandler';
+
 export default function ReportModal() {
   const [report, setReport] = useState<Report>({
     category: '',
     content: '',
   });
-  const setModal = useSetRecoilState(modalState);
-  const setError = useSetRecoilState(errorState);
-  const resetOpenMenuBar = useResetRecoilState(openMenuBarState);
   const reportCategory = [
     { id: 'CHEERS', label: '응원의 한 마디' },
     { id: 'OPINION', label: '의견 제시' },
@@ -26,10 +20,6 @@ export default function ReportModal() {
     { id: 'BUG', label: '버그 신고' },
     { id: 'ETC', label: '기타' },
   ];
-  const reportResponse: { [key: string]: string } = {
-    SUCCESS: '의견 주셔서 감사합니다 ❤️',
-    REJECT: '마음을 담아 의견을 보내주세요 ❤️',
-  };
 
   const inputChangeHandler = ({
     target: { name, value },
@@ -42,20 +32,7 @@ export default function ReportModal() {
     }));
   };
 
-  const reportHandler = async () => {
-    if (report.category && report.content.replace(/^\s+|\s+$/g, '')) {
-      try {
-        await instance.post('/pingpong/feedback', report);
-        setModal({ modalName: null });
-        resetOpenMenuBar();
-        alert(reportResponse.SUCCESS);
-      } catch (e) {
-        setError('JH06');
-      }
-    } else {
-      alert(reportResponse.REJECT);
-    }
-  };
+  const reportHandler = useReportHandler(report);
 
   return (
     <div className={styles.container}>
